Move local storage models to the separate folder
